requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
stripe==7.5.0
Pillow==10.1.0
python-magic-bin==0.4.14
django-imagekit==4.1.0
django-storages==1.14.2
boto3==1.29.3
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2 